Fix PHP notices

+++ applets/subscription/twiml.php
@@ -2,18 +2,20 @@
 $ci =& get_instance();
 $list = AppletInstance::getValue('list');
 $action = AppletInstance::getValue('action');
-$number = normalize_phone_to_E164($_REQUEST['From']);
 
-if('add'==$action){
-	if(!$ci->db->query(sprintf('SELECT id FROM subscribers WHERE list=%d AND value=%s',$list,$number))->num_rows())
-		$ci->db->insert('subscribers', array(
-			'list' => $list,
-			'value' => $number,
-			'joined' => time()
-		));
+if(!empty($_REQUEST['From'])) {
+	$number = normalize_phone_to_E164($_REQUEST['From']);
+	if('add' == $action){
+		if(!$ci->db->query(sprintf('SELECT id FROM subscribers WHERE list = %d AND value = %s', $list, $number))->num_rows())
+			$ci->db->insert('subscribers', array(
+				'list' => $list,
+				'value' => $number,
+				'joined' => time()
+			));
+	}
+	else
+		$ci->db->delete('subscribers', array('list' => $list, 'value' => $number));
 }
-else
-	$ci->db->delete('subscribers', array('list' => $list, 'value' => $number));
 
 $response = new Response();
